Module rexpect::session

Main module of rexpect: start new process and interact with it

Reexports

pub use reader::ReadUntil;

Structs

PtyReplSession
A repl session: e.g. bash or the python shell: You have a prompt where a user inputs commands and the shell executes it and writes some output

PtySession
Interact with a process with read/write/signals, etc.

Functions

spawn
Start command in background in a pty session (pty fork) and return a struct with writer and buffered reader (for unblocking reads).

spawn_bash
Spawn bash in a pty session, run programs and expect output

spawn_command
See spawn

spawn_python