Struct rexpect::session::PtyReplSession

pub struct PtyReplSession { /* fields omitted */ }

A repl session: e.g. bash or the python shell: You have a prompt where a user inputs commands and the shell executes it and writes some output

Methods

impl PtyReplSession

pub fn wait_for_prompt(&mut self) -> Result<String>

pub fn execute(&mut self, cmd: &str, ready_regex: &str) -> Result<()>

Send cmd to repl and: 1. wait for the cmd to be echoed (if echo_on=true) 2. wait for the ready string being present

Q: Why can't I just do send_line and immediately continue? A: Executing a command in e.g. bash causes a fork. If the Unix kernel chooses the parent process (bash) to go first and the bash process sends e.g. Ctrl-C then the Ctrl-C goes to nirvana. The only way to prevent this situation is to wait for a ready string being present in the output.

Another safe way to tackle this problem is to use send_line() and wait_for_prompt()

Example:

use rexpect::spawn_bash;

let mut p = spawn_bash(Some(1000))?;
p.execute("cat <(echo ready) -", "ready")?;
p.send_line("hans")?;
p.exp_string("hans")?;

pub fn send_line(&mut self, line: &str) -> Result<usize>

send line to repl (and flush output) and then, if echo_on=true wait for the input to appear. Return: number of bytes written