Struct rexpect::session::PtySession

pub struct PtySession {
    pub process: PtyProcess,
    pub writer: LineWriter<File>,
    pub reader: NBReader,
    pub commandname: String,
}

Interact with a process with read/write/signals, etc.

Fields
process: PtyProcess writer: LineWriter<File> reader: NBReader commandname: String
Methods

impl PtySession

Start a process in a tty session, write and read from it

Example

use rexpect::spawn;

let mut s = spawn("cat", None)?;
s.send_line("hello, polly!")?;
let line = s.read_line()?;
assert_eq!("hello, polly!", line);

fn send_line(&mut self, line: &str) -> Result<usize>

sends string and a newline to process

this is guaranteed to be flushed to the process returns number of written bytes

fn send(&mut self, s: &str) -> Result<usize>

Send string to process. As stdin of the process is most likely buffered, you'd need to call flush() after send() to make the process actually see your input.

Returns number of written bytes

fn send_control(&mut self, c: char) -> Result<()>

Send a control code to the running process and consume resulting output line (which is empty because echo is off)

E.g. send_control('c') sends ctrl-c. Upper/smaller case does not matter.

fn flush(&mut self) -> Result<()>

Make sure all bytes written via send() are sent to the process

fn read_line(&mut self) -> Result<String>

Read one line (blocking!) and return line without the newline (waits until \n is in the output fetches the line and removes \r at the end if present)

fn try_read(&mut self) -> Option<char>

Return Some(c) if a char is ready in the stdout stream of the process, return None otherwise. This is nonblocking.

fn exp_eof(&mut self) -> Result<String>

Wait until we see EOF (i.e. child process has terminated) Return all the yet unread output

fn exp_regex(&mut self, regex: &str) -> Result<(String, String)>

Wait until provided regex is seen on stdout of child process. Return a tuple: 1. the yet unread output 2. the matched regex

fn exp_string(&mut self, needle: &str) -> Result<String>

Wait until provided string is seen on stdout of child process. Return the yet unread output (without the matched string)

fn exp_char(&mut self, needle: char) -> Result<String>

Wait until provided char is seen on stdout of child process. Return the yet unread output (without the matched char)

fn exp_any(&mut self, needles: Vec<ReadUntil>) -> Result<(String, String)>

Wait until any of the provided needles is found.

Return a tuple with: 1. the yet unread string, without the matching needle (empty in case of EOF and NBytes) 2. the matched string

Example:

use rexpect::{spawn, ReadUntil};

let mut s = spawn("cat", None)?;
s.send_line("hello, polly!")?;
s.exp_any(vec![ReadUntil::String("hello".into()),
               ReadUntil::EOF])?;